Tutorial State Finished
Tutorial state is finished and hooked up, and I tested it and it works.
Now we just need to write some content for the tutorial and we'll be all
set.  I'll try to be there as early as I can tomorrow so we can get this
done and submitted early.